Refactor Store
Redux store refactored in preparation for using the schemas.
