+++ src/components/LedgerTable/DeleteButtonCell.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { useDispatch } from 'react-redux';
 import { useParams } from 'react-router-dom';
 
-import { removeTransaction } from '@/store/transactionsSlice';
+import { removeTransaction } from '@/store/transactions';
 
 import DeleteIcon from '@mui/icons-material/Delete';
