+++ src/components/LedgerTable/StatementSeparatorRow.jsx
@@ -2,9 +2,7 @@ import { TableCell, TableRow, Typography } from '@mui/material';
 import PropTypes from 'prop-types';
 
 export default function StatementSeparatorRow({ statementDate }) {
-  console.log('statement date', statementDate);
   if (statementDate === 'Invalid Date') {
-    console.log('invalid date');
     return null;
   }
   return (
