+++ src/components/RepeatedTransactionsModal/ModalDialog.jsx
@@ -11,7 +11,7 @@ import { useEffect, useState } from 'react';
 import { useDispatch } from 'react-redux';
 import { useParams } from 'react-router-dom';
 
-import { createRepeatTransaction } from '@/store/transactionsSlice';
+import { createRepeatTransaction } from '@/store/transactions';
 import AmountField from './AmountField';
 import DescriptionField from './DescriptionField';
 import FrequencyField from './FrequencyField';
