+++ src/store/accounts/constants.js
@@ -1,0 +1,5 @@
+export const AccountType = Object.freeze({
+  SAVINGS: 'Savings',
+  CHECKING: 'Checking',
+  CREDIT_CARD: 'Credit Card',
+});
