+++ src/store/accounts/selectors.js
@@ -1,0 +1,4 @@
+export const selectAccountById = (id) => (state) =>
+  state.accounts.find((account) => account.id === id);
+
+export const selectAccounts = (state) => state.accounts;
