+++ src/store/index.js
@@ -1,9 +1,1 @@
-import { configureStore } from '@reduxjs/toolkit';
-
-import accountsReducer from './accountsSlice';
-
-export default configureStore({
-  reducer: {
-    accounts: accountsReducer,
-  },
-});
+export { default } from './store';
