+++ src/store/rootReducer.js
@@ -1,0 +1,7 @@
+import { combineReducers } from '@reduxjs/toolkit';
+
+import accountsReducer from './accounts';
+
+export default combineReducers({
+  accounts: accountsReducer,
+});
