+++ src/store/store.js
@@ -1,0 +1,7 @@
+import { configureStore } from '@reduxjs/toolkit';
+
+import rootReducer from './rootReducer';
+
+export default configureStore({
+  reducer: rootReducer,
+});
