+++ src/store/transactions/constants.js
@@ -1,0 +1,6 @@
+export const TransactionStatusEnum = Object.freeze({
+  PENDING: 'pending ',
+  COMPLETE: 'complete ',
+  SCHEDULED: 'scheduled ',
+  PLANNED: 'planned ',
+});
