+++ src/store/transactions/generators.js
@@ -1,0 +1,13 @@
+export const generateTransactionObject = (
+  id,
+  status,
+  date,
+  amount,
+  description
+) => ({
+  id,
+  status,
+  date,
+  amount,
+  description,
+});
