+++ src/views/AccountLedger/AccountNameEdit.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { useState } from 'react';
 import { useDispatch } from 'react-redux';
 
-import { editAccountName } from '@/store/accountsSlice';
+import { editAccountName } from '@/store/accounts';
 
 import { Cancel as CancelIcon, Check as SaveIcon } from '@mui/icons-material';
