+++ src/views/Accounts/CreateNewButton.jsx
@@ -1,7 +1,7 @@
 import { Button } from '@mui/material';
 import { useDispatch } from 'react-redux';
 
-import { createNewAccount } from '@/store/accountsSlice';
+import { createNewAccount } from '@/store/accounts';
 
 export default function CreateNewButton() {
   const dispatch = useDispatch();
